LPD-67439 Follow the standard for Object Actions and Rename it
